+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Release date: work in progress 🚧
 * Adding multiple GPX files at once
 * Adding an animation to smoothly fade photos in and out
 * New error dialog with additional internal information for better error reports
+* New and clear error message on missing track data in GPX file
 * Fixing `NullPointerException` when starting GPX Animator the first time
 * Fixing `NullPointerException` when showing the Changelog
 * Migrating to Java 15